+++ R/pub_hlth_status_by_cnty.R
@@ -1,0 +1,28 @@
+# Functions Roxygen format
+#' @title COVID-19 Public Health Status by County from ESRI
+#'
+#' @description
+#'   This functions pulls the public health status by county in the US as of the
+#'   current date.  The data comes from ESRI open data website.  The data is 1 row per
+#'   county.
+#'   source: https://coronavirus-disasterresponse.hub.arcgis.com/datasets/97792521be744711a291d10ecef33a61/data?geometry=76.921%2C-16.701%2C-109.056%2C72.161
+#'
+#' @return A datarame/tibble
+#'
+#' @examples
+#' \dontrun{
+#' pub_status_cnty <- R.COVID.19::pub_hlth_status_by_cnty()
+#' }
+#'
+#' @importFrom  magrittr %>%
+#' @importFrom  readr read_csv
+#'
+#' @export
+
+
+pub_hlth_status_by_cnty <- function() {
+  # Get data
+  url_csv <- "https://opendata.arcgis.com/datasets/97792521be744711a291d10ecef33a61_0.csv"
+  csv_df <- readr::read_csv(url(url_csv))
+  return(invisible(csv_df))
+}