+++ pkg_dev/test_functions.R
@@ -30,4 +30,25 @@ ggplot(tarr, aes(fill = estimate, color = estimate)) +
 vt <- get_acs(geography = "county", variables = "B19013_001", state = "VT")
 
 
+# ESRI Open Data
+## COVID19_Public_Health_Status_by_County
+## source: https://coronavirus-disasterresponse.hub.arcgis.com/datasets/97792521be744711a291d10ecef33a61/data?geometry=76.921%2C-16.701%2C-109.056%2C72.161
+
+### The Data
+url_csv <- "https://opendata.arcgis.com/datasets/97792521be744711a291d10ecef33a61_0.csv"
+pub_hlth_status_by_cnty <- readr::read_csv(url(url_csv))
+unique(pub_hlth_status_by_cnty$notes)
+
+pub_hlth_status_by_cnty %>% filter(state_name == "Minnesota") %>% View()
+### Shape file
+u_shp <- "https://opendata.arcgis.com/datasets/97792521be744711a291d10ecef33a61_0.zip"
+download.file(u_shp, "pub_hlth_status_by_cnty.zip")
+unzip("pub_hlth_status_by_cnty.zip")
+
+pub_hlth_status_by_cnty_shp <- sf::st_read("COVID19_Public_Health_Emergency_Status_by_County.shp")
+
+pub_status_cnty <- R.COVID.19::pub_hlth_status_by_cnty()
+
+pub_status_cnty_shp <- R.COVID.19::pub_hlth_status_by_cnty_shp()
+